requirements.txt
Pillow~=9.4.0
PyInstaller==5.9.0
requests~=2.32.3